requirements.txt
psycopg2-binary
-dotenv
tkcalendar   # optional if you want date widgets

